+++ packages/celest_cloud/lib/src/cloud/operations/operations.dart
@@ -1,6 +1,7 @@
 import 'package:celest_cloud/src/cloud/base/base_service.dart';
 import 'package:celest_cloud/src/cloud/operations/operations_protocol.dart';
 import 'package:celest_cloud/src/grpc.dart';
+import 'package:celest_cloud/src/util/well_known.dart';
 import 'package:logging/logging.dart';
 
 final class Operations with BaseService {
@@ -36,5 +37,16 @@ final class Operations with BaseService {
   Future<void> cancel(String name) async {
     final request = CancelOperationRequest(name: name);
     return run('Operations.Cancel', request: request, action: _protocol.cancel);
+  }
+
+  Future<Operation> wait(
+    String name, {
+    Duration? timeout,
+  }) async {
+    final request = WaitOperationRequest(
+      name: name,
+      timeout: timeout?.toProto(),
+    );
+    return run('Operations.Wait', request: request, action: _protocol.wait);
   }
 }